Added Numpy-based implementation of the stencils.
* All components feature Numpy-based implementation alongside gt-powered implementation.
* Tests have been updated accordingly.
* TODO: inlined documentation.
* TODO: refresh drivers.

+++ tasmania/python/framework/parallel_splitting.py
@@ -277,7 +277,7 @@ class ParallelSplitting:
                 integrator_gt_powered = process.get("gt_powered", gt_powered)
                 integrator_kwargs = process.get(
                     "time_integrator_kwargs",
-                    {"backend": "numpy", "dtype": np.float32, "rebuild": False},
+                    {"backend": "numpy", "dtype": np.float64, "rebuild": False},
                 )
 
                 self._component_list.append(
